requirements.txt
beautifulsoup4==4.14.3
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
rich==14.2.0
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
